+++ lib/src/ffi/xhvp_library.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:ffi/ffi.dart';
 
+import 'init_timing.dart';
 import 'xhvp_bindings.dart';
 
 /// Loads the native player library and exposes [XhvpBindings].
@@ -19,19 +20,21 @@ class XhvpLibrary {
   static bool get isInitialized => _instance != null;
 
   static XhvpBindings _open() {
-    final DynamicLibrary dylib;
-    if (Platform.isMacOS || Platform.isIOS) {
-      dylib = DynamicLibrary.process();
-    } else if (Platform.isAndroid) {
-      dylib = DynamicLibrary.open('libxue_hua_video_player.so');
-    } else if (Platform.isWindows) {
-      dylib = DynamicLibrary.open('xue_hua_video_player_plugin.dll');
-    } else if (Platform.isLinux) {
-      dylib = DynamicLibrary.open('libxue_hua_video_player_plugin.so');
-    } else {
-      throw UnsupportedError('Unsupported platform for XhvpLibrary');
-    }
-    return XhvpBindings(dylib);
+    return xhvpTimed('dylib_open', () {
+      final DynamicLibrary dylib;
+      if (Platform.isMacOS || Platform.isIOS) {
+        dylib = DynamicLibrary.process();
+      } else if (Platform.isAndroid) {
+        dylib = DynamicLibrary.open('libxue_hua_video_player.so');
+      } else if (Platform.isWindows) {
+        dylib = DynamicLibrary.open('xue_hua_video_player_plugin.dll');
+      } else if (Platform.isLinux) {
+        dylib = DynamicLibrary.open('libxue_hua_video_player_plugin.so');
+      } else {
+        throw UnsupportedError('Unsupported platform for XhvpLibrary');
+      }
+      return XhvpBindings(dylib);
+    });
   }
 
   /// For host unit tests that load a built dylib by path.